# Adds Benchmark for XSS detection, minor refactor/optimization