--- a/html5.go
+++ b/html5.go
@@ -73,11 +73,12 @@ func (h *h5State) stateBogusComment2() bool {
 // 12.2.4.49
 // 12.2.4.50
 // 12.2.4.51
-//   state machine spec is confusing since it can only look
-//   at one character at a time but simply it's comments end by:
-//   1) EOF
-//   2) ending in -->
-//   3) ending in -!>
+//
+//	state machine spec is confusing since it can only look
+//	at one character at a time but simply it's comments end by:
+//	1) EOF
+//	2) ending in -->
+//	3) ending in -!>
 func (h *h5State) stateComment() bool {
 	pos := h.pos
 
@@ -92,14 +93,14 @@ func (h *h5State) stateComment() bool {
 			h.tokenType = html5TypeTagComment
 			return true
 		}
-		offset := 1
-
+		offset := 1 // offset is set at the beginning of each iteration
+		posIndexOffset := pos + index + offset
 		// skip all nulls
-		for pos+index+offset < h.len && h.s[pos+index+offset] == 0x00 {
-			offset++
+		for posIndexOffset < h.len && h.s[posIndexOffset] == 0x00 {
+			posIndexOffset++ //offset++
 		}
 
-		if pos+index+offset == h.len {
+		if posIndexOffset == h.len {
 			h.state = h.stateEOF
 			h.tokenStart = h.s[h.pos:]
 			h.tokenLen = h.len - h.pos
@@ -107,14 +108,15 @@ func (h *h5State) stateComment() bool {
 			return true
 		}
 
-		ch := h.s[pos+index+offset]
+		ch := h.s[posIndexOffset]
 		if ch != byteDash && ch != byteBang {
 			pos = pos + index + 1
+			// pos updated, but the next iteration starts and posIndexOffset is calculated again
 			continue
 		}
-		offset++
+		posIndexOffset++ //offset++
 
-		if pos+index+offset == h.len {
+		if posIndexOffset == h.len {
 			h.state = h.stateEOF
 			h.tokenStart = h.s[h.pos:]
 			h.tokenLen = h.len - h.pos
@@ -122,16 +124,17 @@ func (h *h5State) stateComment() bool {
 			return true
 		}
 
-		if h.s[pos+index+offset] != byteGT {
+		if h.s[posIndexOffset] != byteGT {
 			pos = pos + index + 1
+			// pos updated, but the next iteration starts and posIndexOffset is calculated again
 			continue
 		}
-		offset++
+		posIndexOffset++ //offset++
 
 		// ends in --> or -!>
 		h.tokenStart = h.s[h.pos:]
 		h.tokenLen = index + pos - h.pos
-		h.pos = pos + index + offset
+		h.pos = posIndexOffset
 		h.state = h.stateData
 		h.tokenType = html5TypeTagComment
 		return true