--- a/sqli_test.go
+++ b/sqli_test.go
@@ -147,19 +147,19 @@ func TestSQLiDriver(t *testing.T) {
 		data := readTestData(p)
 		switch {
 		case strings.Contains(fi.Name(), "-sqli-"):
-			t.Run("SQLi Driver - sqli", func(t *testing.T) {
+			t.Run(fi.Name(), func(t *testing.T) {
 				runSQLiTest(t, data, p, fingerprints, 0)
 			})
 		case strings.Contains(fi.Name(), "-folding-"):
-			t.Run("SQLi Driver - folding", func(t *testing.T) {
+			t.Run(fi.Name(), func(t *testing.T) {
 				runSQLiTest(t, data, p, folding, sqliFlagQuoteNone|sqliFlagSQLAnsi)
 			})
 		case strings.Contains(fi.Name(), "-tokens_mysql-"):
-			t.Run("SQLi Driver - tokens_mysql", func(t *testing.T) {
+			t.Run(fi.Name(), func(t *testing.T) {
 				runSQLiTest(t, data, p, tokens, sqliFlagQuoteNone|sqliFlagSQLMysql)
 			})
 		case strings.Contains(fi.Name(), "-tokens-"):
-			t.Run("SQLi Driver - tokens", func(t *testing.T) {
+			t.Run(fi.Name(), func(t *testing.T) {
 				runSQLiTest(t, data, p, tokens, sqliFlagQuoteNone|sqliFlagSQLAnsi)
 			})
 		}