--- a/tests/test-html5-068.txt
+++ b/tests/test-html5-068.txt
@@ -1,0 +1,12 @@
+--TEST--
+tags
+--INPUT--
+quitesomerandomcharacters<TABLE BACKGROUND="javascript:javascript:alert(99111)">
+--EXPECTED--
+DATA_TEXT,25,quitesomerandomcharacters
+TAG_NAME_OPEN,5,TABLE
+ATTR_NAME,10,BACKGROUND
+ATTR_VALUE,34,javascript:javascript:alert(99111)
+TAG_NAME_CLOSE,1,>
+
+