--- a/tests/test-html5-070.txt
+++ b/tests/test-html5-070.txt
@@ -1,0 +1,10 @@
+--TEST--
+Extended bogus closing tags
+--INPUT--
+</ looooooonger foo="><script>ReferenceError.prototype.__defineGetter__('name', function(){javascript:alert(1)}),x</script>
+--EXPECTED--
+TAG_COMMENT,19, looooooonger foo="
+TAG_NAME_OPEN,6,script
+TAG_NAME_CLOSE,1,>
+DATA_TEXT,84,ReferenceError.prototype.__defineGetter__('name', function(){javascript:alert(1)}),x
+TAG_CLOSE,6,script