This PR:
- Adds a Benchmark test for XSS detection
- Reduces functions called inside interactions (still no performance improvements, I think mainly because of the small amount of data iterated)
- Has minor refactors in favor of Go idioms rather than C ones. 

Benchmark Before/After:
```
name                  old time/op    new time/op   delta
XSSDriver/html5-10    42.5µs ± 3%    41.2µs ± 1%  -3.19%  (p=0.000 n=100+86)
```