@anuraaga 2022-09-20
> XSSDriver/html5-10    42.5µs ± 3%    41.2µs ± 1%  -3.19%  (p=0.000 n=100+86)

Can you add some logic to check for, or at least debug that the benchmark is actually running code? Even if the inputs are all small, there are about 60 test case files meaning <1 microsecond per case, this seems too low. Nothing seems obviously wrong with the code from reading it though 🤔