@fzipi 2023-04-02
@M4tteoP Can you push some additional larger texts for https://github.com/corazawaf/libinjection-go/pull/13#pullrequestreview-1113024471 ?