@M4tteoP 2024-01-27
Requires https://github.com/corazawaf/libinjection-go/pull/18