@M4tteoP 2024-06-05
If you are happy with the additional larger texts added, it is ready to be merged from my side